Extract data compilation into ExportService
