+++ app/Http/Controllers/ExportController.php
@@ -3,25 +3,26 @@
 namespace Rogue\Http\Controllers;
 
 use Rogue\Models\Signup;
-use Rogue\Services\Registrar;
+use Rogue\Services\ExportService;
 
 class ExportController extends Controller
 {
     /**
-     * Registrar instance
+     * ExportService instance
      *
-     * @var Rogue\Services\Registrar
+     * @var Rogue\Services\ExportService
      */
-    protected $registrar;
+    protected $export;
 
     /**
      * Instantiate a new ExportController instance.
      *
      * @param Rogue\Services\Registrar $registrar
      */
-    public function __construct(Registrar $registrar)
+    public function __construct(ExportService $export)
     {
-        $this->registrar = $registrar;
+        $this->middleware('auth');
+        $this->export = $export;
     }
 
     /**
@@ -35,23 +36,7 @@ class ExportController extends Controller
         $signups = Signup::whereNotNull('details')->where('campaign_id', $campaignId)->get();
 
         // Compile the data
-        $final_results = [];
-
-        foreach ($signups as $signup) {
-            $northstar_user = $this->registrar->find($signup->northstar_id);
-
-            $next_row = [
-                'campaign_id' => $signup->campaign_id,
-                'campaign_run_id' => $signup->campaign_run_id,
-                'northstar_id' => $signup->northstar_id,
-                'first_name' => $northstar_user->first_name,
-                'email' => $northstar_user->email,
-                'mobile' => $northstar_user->mobile,
-                'age' => getAgeFromBirthdate($northstar_user->birthdate),
-            ];
-
-            array_push($final_results, $next_row);
-        }
+        $final_results = $this->export->exportSignups($signups);
 
         // Format as csv
         $output = '';
